refactor: reorganize PegStock application structure
