+++ src/main/java/com/pegazuls/aerodesign/PegStock/PegStockApplication.java
@@ -4,10 +4,12 @@ import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.context.annotation.ComponentScan;
 
+
 @SpringBootApplication
-@ComponentScan(basePackages = {"com.pegazuls.aerodesign.PegStock", "front.controllers"})
+@ComponentScan(basePackages = {"com.pegazuls.aerodesign.PegStock"})
 public class PegStockApplication {
-    public static void main(String[] args) {
-        SpringApplication.run(PegStockApplication.class, args);
+
+        public static void main(String[] args) {
+            SpringApplication.run(PegStockApplication.class, args);
     }
 }
